+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 3.3.4
 ----------------
 - Fix "Create User Account" schools in User.php
+- Security check for $modname in Modules.php
 
 Changes in 3.3.3
 ----------------